arm64: dts: allwinner: A100: Add RTC nodes
The RTC serves two purposes on the A100; it keeps track of time, but it
also manages two clock sources: the 16MHz RC oscillator (iosc) and an
optional 32 KHz external crystal (losc/osc32k). Additionally, it can
provide an additional 32 KHz clock output (losc_fanout) to external
devices.

On reset, the clock gate is open; thus it was possible to represent iosc
and losc using fixed clocks. However, by doing so, we lose the ability
to use the fanout, which on some devices is connected to an SDIO
wireless adapter.

Now that we have an RTC driver, add the RTC node to the device tree, and
remove the previous nodes representing iosc and osc32k. For all
peripherals that referenced these clocks before, reference the clocks on
the RTC node instead.

Signed-off-by: Cody Eksal <masterr3c0rd@epochal.quest>