net: stmmac: dwmac-sun8i: Add support for the A100
The A100 has a similar PHY to other Allwinner SoCs. On reset, the syscon
register's reset value is 0x58000; this would match the H3's version,
except that there is no internal PHY. Additionally, the datasheet claims
MII is not supported, only RMII and RGMII. Add a new variant for the
A100.

Signed-off-by: Cody Eksal <masterr3c0rd@epochal.quest>