arm64: dts: allwinner: A100: Add ethernet, syscon
The A100 has an ethernet controller similar to those found on other
Allwinner SoCs Additionally, it has a system configuration block
(SYS_CFG in the datasheet's memory map), and three SRAM regions: A1, A2,
and C. The SYS_CFG region is not documented in the user manual, but a
glance at the vendor tree points to it being closest to the A64. Add
all of these to the base device tree.

Signed-off-by: Cody Eksal <masterr3c0rd@epochal.quest>